+++ packages/vite/client.d.ts
@@ -1,5 +1,5 @@
 /// <reference lib="dom" />
-/// <reference path="./types/importMeta" />
+/// <reference path="./types/importMeta.d.ts" />
 
 // CSS modules
 type CSSModuleClasses = { readonly [key: string]: string }
@@ -95,6 +95,10 @@ declare module '*.webp' {
   const src: string
   export default src
 }
+declare module '*.avif' {
+  const src: string
+  export default src
+}
 
 // media
 declare module '*.mp4' {
@@ -148,6 +152,16 @@ declare module '*.otf' {
   export default src
 }
 
+// other
+declare module '*.wasm' {
+  const src: string
+  export default src
+}
+declare module '*.webmanifest' {
+  const src: string
+  export default src
+}
+
 // web worker
 declare module '*?worker' {
   const workerConstructor: {
@@ -176,6 +190,11 @@ declare module '*?raw' {
 }
 
 declare module '*?url' {
+  const src: string
+  export default src
+}
+
+declare module '*?inline' {
   const src: string
   export default src
 }