+++ packages/vite/src/node/plugins/resolve.ts
@@ -59,6 +59,7 @@ export interface InternalResolveOptions extends ResolveOptions {
   asSrc?: boolean
   tryIndex?: boolean
   tryPrefix?: string
+  skipPackageJson?: boolean
   preferRelative?: boolean
   isRequire?: boolean
 }
@@ -105,7 +106,7 @@ export function resolvePlugin(baseOptions: InternalResolveOptions): Plugin {
 
       const options = isRequire ? requireOptions : baseOptions
 
-      let res
+      let res: string | PartialResolvedId | undefined
 
       // explicit fs paths that starts with /@fs/*
       if (asSrc && id.startsWith(FS_PREFIX)) {
@@ -286,7 +287,8 @@ function tryFsResolve(
       options,
       false,
       targetWeb,
-      options.tryPrefix
+      options.tryPrefix,
+      options.skipPackageJson
     ))
   ) {
     return res
@@ -300,7 +302,8 @@ function tryFsResolve(
         options,
         false,
         targetWeb,
-        options.tryPrefix
+        options.tryPrefix,
+        options.skipPackageJson
       ))
     ) {
       return res
@@ -314,7 +317,8 @@ function tryFsResolve(
       options,
       tryIndex,
       targetWeb,
-      options.tryPrefix
+      options.tryPrefix,
+      options.skipPackageJson
     ))
   ) {
     return res
@@ -327,7 +331,8 @@ function tryResolveFile(
   options: InternalResolveOptions,
   tryIndex: boolean,
   targetWeb: boolean,
-  tryPrefix?: string
+  tryPrefix?: string,
+  skipPackageJson?: boolean
 ): string | undefined {
   let isReadable = false
   try {
@@ -341,11 +346,13 @@ function tryResolveFile(
     if (!fs.statSync(file).isDirectory()) {
       return normalizePath(ensureVolumeInPath(file)) + postfix
     } else if (tryIndex) {
-      const pkgPath = file + '/package.json'
-      if (fs.existsSync(pkgPath)) {
-        // path points to a node package
-        const pkg = loadPackageData(pkgPath)
-        return resolvePackageEntry(file, pkg, options, targetWeb)
+      if (!skipPackageJson) {
+        const pkgPath = file + '/package.json'
+        if (fs.existsSync(pkgPath)) {
+          // path points to a node package
+          const pkg = loadPackageData(pkgPath)
+          return resolvePackageEntry(file, pkg, options, targetWeb)
+        }
       }
       const index = tryFsResolve(file + '/index', options)
       if (index) return index + postfix
@@ -371,7 +378,7 @@ export function tryNodeResolve(
   const deepMatch = id.match(deepImportRE)
   const pkgId = deepMatch ? deepMatch[1] || deepMatch[2] : id
 
-  let basedir
+  let basedir: string
   if (dedupe && dedupe.includes(pkgId)) {
     basedir = root
   } else if (
@@ -501,7 +508,7 @@ function loadPackageData(pkgPath: string, cacheKey = pkgPath) {
   const data = JSON.parse(fs.readFileSync(pkgPath, 'utf-8'))
   const pkgDir = path.dirname(pkgPath)
   const { sideEffects } = data
-  let hasSideEffects
+  let hasSideEffects: (id: string) => boolean
   if (typeof sideEffects === 'boolean') {
     hasSideEffects = () => sideEffects
   } else if (Array.isArray(sideEffects)) {
@@ -604,6 +611,15 @@ export function resolvePackageEntry(
   }
 
   entryPoint = entryPoint || data.main || 'index.js'
+
+  // make sure we don't get scripts when looking for sass
+  if (
+    options.mainFields?.[0] === 'sass' &&
+    !options.extensions?.includes(path.extname(entryPoint))
+  ) {
+    entryPoint = ''
+    options.skipPackageJson = true
+  }
 
   // resolve object browser field in package.json
   const { browser: browserField } = data